requirements.txt
--index-url https://pypi.python.org/simple/
-e ./vendor/geonode
-e ./vendor/django-osgeo-importer
git+https://github.com/boundlessgeo/sphinx@bex#egg=Sphinx
git+https://github.com/venicegeo/nearsight@master#egg=nearsight
-e ./vendor/geonode-elasticsearch
django-jsonfield==1.0.1
django-jsonfield-compat==0.4.4
numpy==1.13.3
django-geonode-client==1.0.0
dj-database-url==0.4.2
celery[redis]==3.1.18
django-storages==1.6.5
boto==2.48.0
waitress==1.1.0
whitenoise==3.3.1
django-cors-headers==1.3.1
django-classification-banner==0.1.5
django-solo==1.1.2
django-colorfield==0.1.14
psycopg2==2.7.3.2
python-ldap==2.5.2
django-auth-ldap==1.3.0
GDAL==2.1.0
supervisor==3.3.3
supervisor-stdout==0.1.1
python-resize-image==1.1.11
django-flat-theme==1.1.4
flake8==3.5.0
elasticsearch-dsl>=6.0.0,<7.0.0
social-auth-app-django==1.2.0
mock==2.0.0
Django==1.8.18
six==1.10.0
djangorestframework==3.6.4
